♻️ Move Duration to opts and add temporary alias to sqlslog package
